requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116/

pandas==2.1.2
matplotlib==3.8.3
sahi==0.11.14
bbox_visualizer==0.1.0
scikit-image==0.21.0
imagecodecs==2023.9.18
albumentations==1.3.1
numpy==1.24.4
opencv_python==4.8.1.78
opencv_python_headless==4.5.5.64
patchify==0.2.3
pika==1.2.0
Pillow==10.0.1
pydantic==1.10.2
six==1.16.0
PyYAML==6.0.1
packaging==23.2
torch==1.13.1+cu116
torchvision==0.14.1+cu116





